requirements.txt
flask==1.1.1
flask-login==0.4.1
pendulum==2.0.5
lxml==4.4.0
python-magic==0.4.15
flask-bcrypt==0.7.1
flask-sqlalchemy==2.4.0
flask-migrate==2.5.2
filemagic==1.6
flask-script==2.0.6
pyjwt==1.4.2
gunicorn
flask-cors
celery==4.2.1
psycopg2==2.8.3
sentry-sdk[flask]==0.11.2
eventlet==0.25.1
python-dotenv==0.10.3
requests==2.22.0
redis==3.2.1
numpy==1.17.3
keras==2.2.5
tensorflow==1.13.1
sklearn